MapObject added
add all the map object.  bomb and boss(can place bomb as well)
implemented. the player can place bomb. the next step is to make the
graphic engine to refresh the pane and the game engine: to move the
listener to the game engine and invoke the up down action.